Subido todo el sistema de administracion de redactores.

+++ momoko-backend/src/main/java/com/momoko/es/backend/model/repository/UsuarioRepository.java
@@ -16,7 +16,7 @@ import com.momoko.es.backend.model.entity.UsuarioEntity;
 /**
  * The Interface UsuarioRepositoryImpl.
  */
-public interface UsuarioRepository extends CrudRepository<UsuarioEntity, Long> {
+public interface UsuarioRepository extends CrudRepository<UsuarioEntity, Integer> {
 
     /**
      * Find by email.
